requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dill==0.3.6
gensim==4.3.0
hdbscan==0.8.29
ipython==8.7.0
joblib==1.2.0
matplotlib==3.6.2
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
SQLAlchemy==2.0.9
streamlit==1.23.1
